requirements.txt
# Dependencias para Home Control en Raspberry Pi

# Django y REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Servidor WSGI para producción
gunicorn==21.2.0

# Archivos estáticos
whitenoise==6.6.0

# Cliente MQTT
paho-mqtt==1.6.1

# Variables de entorno
python-dotenv==1.0.0

# HTTP requests para MQTT bridge
requests==2.31.0

# Timezone support
pytz==2023.3

# Utilidades adicionales (para imágenes si se necesitan)
pillow==10.1.0

# Extensiones de Django (opcional para desarrollo)
django-extensions==3.2.3

# Base de datos PostgreSQL (opcional)
# psycopg2-binary==2.9.9

# Desarrollo y testing (opcional)
# django-debug-toolbar==4.2.0
# django-ratelimit==4.1.0
